feat(agent): 更新smartmontools嵌入方式并添加.gitkeep文件
修改smart_windows.go以使用embed.FS嵌入整个smartmontools目录，并添加验证逻辑
添加.gitkeep文件以满足go:embed模式要求
更新.gitignore以排除不必要的二进制目录
修复smart_test.go中的容量类型声明
删除过时的多语言翻译文档

+++ agent/smart_test.go
@@ -45,7 +45,7 @@ func TestParseSmartForScsi(t *testing.T) {
 	assert.Equal(t, deviceData.DiskType, "scsi")
 	assert.EqualValues(t, deviceData.Temperature, 34)
 	assert.Equal(t, deviceData.SmartStatus, "PASSED")
-	assert.EqualValues(t, deviceData.Capacity, 14000519643136)
+	assert.EqualValues(t, deviceData.Capacity, uint64(14000519643136))
 
 	if len(deviceData.Attributes) == 0 {
 		t.Fatalf("expected attributes to be populated")
